pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musicgen"
version = "0.1.0"
description = "A rule-based orchestral music generation library"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "MusicGen Contributors", email = "contributors@musicgen.dev"}
]
keywords = ["music", "composition", "midi", "music-theory", "orchestration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: MIDI",
    "Topic :: Artistic Software",
]

dependencies = [
    "music21>=9.0",
    "mido>=1.2",
    "pretty-midi>=0.2",
    "numpy>=1.24",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# Core development tools (Python 3.10+)
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=1.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
]

# Audio synthesis (requires system dependencies: fluidsynth)
audio = [
    "pydub>=0.25.0",
]

# LilyPond/PDF export (requires Python <3.12)
lilypond = [
    "abjad>=3.0",
]

# AI-powered composition (Gemini)
ai = [
    "google-genai>=1.0.0",
    "google-api-core>=2.0",
    "python-dotenv>=1.0",
]

# All optional dependencies (except lilypond for Python 3.12+)
all = [
    "musicgen[dev,audio,ai]",
]

[project.urls]
Homepage = "https://github.com/musicgen/music-gen-lib"
Documentation = "https://music-gen-lib.readthedocs.io"
Repository = "https://github.com/musicgen/music-gen-lib"

[project.scripts]
musicgen = "musicgen.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = false
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "mido.*",
    "pretty_midi.*",
    "yaml.*",
    "google.*",
    "genai.*",
    "pydub.*",
    "music21.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "W191",   # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"tests/*" = ["ARG"]       # unused arguments in tests
"src/musicgen/__main__.py" = ["F401", "F841"]  # availability checks and unused vars
"src/musicgen/theory/*" = ["ARG", "SIM", "F841", "B007"]  # theory module patterns
"src/musicgen/ai/*" = ["ARG", "B904"]  # ai module patterns
"src/musicgen/ai_client/*" = ["ARG", "B904"]  # ai client patterns
"src/musicgen/composer_new/*" = ["B904"]  # composer patterns
"src/musicgen/ai_models/*" = ["UP"]  # using Union for compatibility
"src/musicgen/io/*" = ["ARG", "B904", "F841"]  # io module patterns
"src/musicgen/orchestration/*" = ["ARG", "F841", "F821"]  # orchestration patterns
"src/musicgen/renderer/*" = ["ARG", "B904", "F401"]  # renderer patterns

[tool.ruff.lint.isort]
known-first-party = ["musicgen"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
